Refactor - replaced redundant crews abilities with shared CrewsVehicleSourceTriggeredAbility
