+++ Mage.Sets/src/mage/sets/kaladesh/SpeedwayFanatic.java
@@ -29,18 +29,13 @@ package mage.sets.kaladesh;
 
 import java.util.UUID;
 import mage.MageInt;
-import mage.abilities.TriggeredAbilityImpl;
-import mage.abilities.effects.Effect;
+import mage.abilities.effects.common.CrewsVehicleSourceTriggeredAbility;
 import mage.abilities.effects.common.continuous.GainAbilityTargetEffect;
 import mage.abilities.keyword.HasteAbility;
 import mage.cards.CardImpl;
 import mage.constants.CardType;
 import mage.constants.Duration;
 import mage.constants.Rarity;
-import mage.constants.Zone;
-import mage.game.Game;
-import mage.game.events.GameEvent;
-import mage.target.targetpointer.FixedTarget;
 
 /**
  *
@@ -74,40 +69,3 @@ public class SpeedwayFanatic extends CardImpl {
     }
 }
 
-class CrewsVehicleSourceTriggeredAbility extends TriggeredAbilityImpl {
-
-    public CrewsVehicleSourceTriggeredAbility(Effect effect) {
-        super(Zone.BATTLEFIELD, effect, false);
-    }
-
-    public CrewsVehicleSourceTriggeredAbility(final CrewsVehicleSourceTriggeredAbility ability) {
-        super(ability);
-    }
-
-    @Override
-    public CrewsVehicleSourceTriggeredAbility copy() {
-        return new CrewsVehicleSourceTriggeredAbility(this);
-    }
-
-    @Override
-    public boolean checkEventType(GameEvent event, Game game) {
-        return event.getType() == GameEvent.EventType.CREWED_VEHICLE;
-    }
-
-    @Override
-    public boolean checkTrigger(GameEvent event, Game game) {
-        if (event.getTargetId().equals(getSourceId())) {
-            for (Effect effect : getEffects()) {
-                // set the vehicle id as target
-                effect.setTargetPointer(new FixedTarget(event.getSourceId()));
-            }
-            return true;
-        }
-        return false;
-    }
-
-    @Override
-    public String getRule() {
-        return "When {this} crews a Vehicle, " + super.getRule();
-    }
-}
